maps: moving the sdk into a versioned folder
