+++ internal/services/maps/sdk/2021-02-01/accounts/method_regeneratekeys_autorest.go
